#620: Fixed minor blunder.
git-svn-id: svn://dev.exiv2.org/svn@1885 b7c8b350-86e7-0310-a4b4-de8f6a8f16a3

+++ branches/unstable/src/tiffcomposite_int.hpp
@@ -1243,7 +1243,7 @@ namespace Exiv2 {
          */
         uint32_t tagStep() const { return elDefaultDef_.size(0, group_); }
         //DATA
-        uint32_t    group_;        //!< Group for the elements
+        uint16_t    group_;        //!< Group for the elements
         ByteOrder   byteOrder_;    //!< Byte order, invalidByteOrder to inherit
         TiffType    elTiffType_;   //!< Type for the array entry and the size element, if any
         CryptFct    cryptFct_;     //!< Crypt function, 0 if not used
